Fixed the different cards after launching second game bug. However there is still the incorrect guess bug and now images don't want to show
